Initial researcher app
Added a researcher app to the project, altering the base settings file
to include the aforementioned app.  Researcher model basic structure
started.  Used a reference into the auth table instead of actually
keeping username, id, password in researcher, as we can harness the
existing auth models for this.  Only real interesting thing in this
model is the authLevel that will provide some granularity to researcher
statuses.