Admin reskinning
Took a bit to realign to this merge.  Proceeded with adding permissions
to researcher, added researcher as a custom user manager (may not be the
most exact terminology, but that's what the settings thing
AUTH_PROFIL... does).  Added some admin override templates, these will
override the basic activity of the admin site.  Next step is to
templatize (good verb?) the mockups for the researcher view.  Anything
broken that I haven't checked is Ken's fault, I'm sure blame will prove
it!