Django basic researcher skeleton
Using the Django admin framework, I made a skeleton of the researcher
view.  Work is to be done on fleshing out the action delegation, and
some template work on making the Django admin interface look like the
mockup.  Stubbed in a template model, but it should probably contain a
short description.  Running into some issues with the single app
version, as you can't as easily override admin templates for each app.