templates moved, website logic updated
Moved my html templates to the templates folder: It's specified in the
settings.py but wasn't in the repository because it was empty. Also
changed the urls.py to include student pages (PLACEHOLDER), and changed
views.py to do some error checking.
Also, including my rough python script for filling the databases. not
tested yet, need to verify what exactly goes in each database.