Further dev on template wizard
I think I've made significant progress, but I'm staring at some session
variable juggling issues.  Trying to map out a total conversation in
session variables seems to be a bit nontrivial, I have one
video/response working, but extending it to the chain of videos needs a
little thought.  Likely it'll require plugging the working video
instance+responses etc into a list labeled with the video it relates to.
I haven't touched on the database stuff yet, as that'll be trivial next
to mapping the thing out in session.  Everything is rough, especially
variable consistency.  I've located a rich text plugin, but I haven't
vetted/installed it yet.

You'll notice that the url is /admin/template-wizard.