added css folder, updated Student Login
For sake of simplicity, the user agreement is on the same page as the
Student Login. I also added the css folder to the templates folder.
note: the settings.py for simcom lists "django-templates" as the
template folder, but there is another directory, "templates", which also
has html files. we should pick one folder and stick to it, rather than
have two copies.