Updated GenerateLink, ShareTemplate, & ShareResponse
Added comments to the ShareTemplate & ShareResponse views.

Updated the GenerateLink view to insure that the validationKey is
unique (also made the field ‘validationKey’ in the StudentAccess model
unique).   Also added a error handler to insure that if a URL passed
templateID does not exist, that it will display the page with no
template selected.

Updated the ShareTemplate view.  The old version of the ShareTemplate
view did not copy the  TemplateResponseRel.  Added the code to copy the
TemplateResponseRel.  Also the model for the Template had changed
(firstIstanceID now references the pageInstanceID) so I corrected the
code to reference the pageinstanceID instead of a TemplateFlowRel.
Also added a error handler to insure that if a URL passed templateID
does not exist, that it will display the page with no template selected.

Updated the ShareResponse view to share a conversation instead of just
a single response.

Also fixed some minor syntax errors in order to compile.