Fixed a couple of bugs in student view
Received an error when taking a conversation.  The variable playback
was not setup prior to being called.

Also when at the end of a conversation, the student could refresh the
page and get back to the last response screen.  The student then could
select a different response and it would record it as well.  To fix
this problem, I added a line to flush the session variables before
calling the Student_Submission.html.