Fixed issues with template wizard
When the requirement for endpoint for all videos in template wizard was
committed out, it created an issue when adding a video with no
responses.  After saving and then editing the template, the video with
no responses would not show up in video list.  Added the code to add
the video to the TemplateFlowRel.  Also had to add some code to load
the template (in edit mode), if a video has no responses.